refactor: factor out file copy function
